--- tests/test_zbrains.py
@@ -1,103 +1,0 @@
-import unittest
-from unittest.mock import patch, MagicMock
-import zbrains
-import pytest
-from unittest.mock import patch, MagicMock
-from zbrains import main
-from functions.utilities import ProcessingException
-
-class TestZBrains(unittest.TestCase):
-
-    @patch('zbrains.os')
-    def test_setenv(self, mock_os):
-        mock_os.path.isfile.return_value = True
-        result = zbrains.setenv("/path/to/workbench")
-        assert result == "/path/to/workbench"
-        mock_os.path.isfile.assert_called_once_with("/path/to/workbench")
-
-    @patch('zbrains.os')
-    def test_setenv_file_not_found(self, mock_os):
-        mock_os.path.isfile.return_value = False
-        with pytest.raises(FileNotFoundError):
-            zbrains.setenv("/path/to/workbench")
-
-    @patch('zbrains.do_cmd')
-    def test_cleanup(self, mock_do_cmd):
-        zbrains.cleanup("/tmp/dir")
-        mock_do_cmd.assert_called_once_with("rm -rf /tmp/dir")
-
-    @patch('zbrains.assert_same_size')
-    @patch('zbrains.assert_required')
-    @patch('zbrains.assert_exists')
-    def test_check_sub(self, mock_exists, mock_required, mock_same_size):
-        args = MagicMock()
-        args.sub = "sub"
-        args.ses = "ses"
-        args.dataset = "dataset"
-        zbrains.check_sub(args, "sub", "ses")
-        mock_exists.assert_called_once()
-        mock_required.assert_called_once()
-        mock_same_size.assert_called_once()
-
-
-    # Mocking external dependencies and environment
-    @pytest.fixture(autouse=True)
-    def mock_dependencies(monkeypatch):
-        monkeypatch.setattr("subprocess.check_output", MagicMock(return_value=b"WorkBench 1.0\n"))
-        monkeypatch.setattr("os.environ", {"WORKBENCH_PATH": "/fake/path"})
-        monkeypatch.setattr("shutil.which", MagicMock(return_value="/usr/bin/wb_command"))
-        monkeypatch.setattr("os.path.isdir", MagicMock(return_value=True))
-        monkeypatch.setattr("os.makedirs", MagicMock())
-        monkeypatch.setattr("tempfile.mkdtemp", MagicMock(return_value="/tmp/fake_dir"))
-        monkeypatch.setattr("os.chmod", MagicMock())
-        monkeypatch.setattr("shutil.rmtree", MagicMock())
-        monkeypatch.setattr("os.path.exists", MagicMock(return_value=True))
-        monkeypatch.setattr("glob.glob", MagicMock(return_value=["/fake/path/surf.json"]))
-        monkeypatch.setattr("os.path.realpath", MagicMock(side_effect=lambda x: x))
-        monkeypatch.setattr("os.path.join", MagicMock(side_effect=lambda *args: "/".join(args)))
-        monkeypatch.setattr("functions.run_proc.run", MagicMock())
-        monkeypatch.setattr("functions.run_analysis.run", MagicMock())
-
-    # Happy path tests with various realistic test values
-    @pytest.mark.parametrize("args,expected", [
-        (MagicMock(run=["proc"], struct=["hippocampus"], feat=["all"], resolution=["all"], label_ctx=["midthickness"], label_hip=["midthickness"], smooth_ctx=2, smooth_hip=2, threshold=0.5, verbose=True, micapipe="micapipe_dir", hippunfold="hippunfold_dir", dataset="dataset_dir", zbrains="zbrains_dir", sub="001", ses="001"), "proc_hippocampus"),
-        (MagicMock(run=["analysis"], struct=["cortex"], feat=["thickness"], resolution=["1mm"], label_ctx=["midthickness"], label_hip=["midthickness"], smooth_ctx=2, smooth_hip=2, threshold=0.5, verbose=True, dataset_ref=["dataset_ref_dir"], zbrains_ref=["zbrains_ref_dir"], demo_ref=["demo_ref_file"], dataset="dataset_dir", zbrains="zbrains_dir", sub="002", ses=None, demo="demo_file", normative=True, deconfound=True, column_map=None), "analysis_cortex"),
-    ], ids=["proc_hippocampus", "analysis_cortex"])
-    def test_main_happy_path(args, expected, mock_dependencies):
-        # Arrange
-
-        # Act
-        main(args)
-
-        # Assert
-        # Assertions are based on the expected behavior of the function, such as creating directories, calling subprocesses, etc.
-        # Since we're mocking, we verify if the mocked functions were called with expected arguments.
-        # This is a simplified example, and actual assertions would depend on the function's side effects and return values.
-
-    # Various edge cases
-    @pytest.mark.parametrize("args,expected_exception", [
-        (MagicMock(run=["unknown_task"]), ProcessingException),
-        (MagicMock(run=["proc"], struct=["unknown_structure"]), ProcessingException),
-    ], ids=["unknown_task", "unknown_structure"])
-    def test_main_edge_cases(args, expected_exception, mock_dependencies):
-        # Arrange
-
-        # Act & Assert
-        with pytest.raises(expected_exception):
-            main(args)
-
-    # Various error cases
-    @pytest.mark.parametrize("args,expected_exception", [
-        (MagicMock(run=["proc"], micapipe=None), ProcessingException),
-        (MagicMock(run=["analysis"], dataset_ref=None), ProcessingException),
-    ], ids=["missing_micapipe", "missing_dataset_ref"])
-    def test_main_error_cases(args, expected_exception, mock_dependencies):
-        # Arrange
-
-        # Act & Assert
-        with pytest.raises(expected_exception):
-            main(args)
-
-
-if __name__ == '__main__':
-    unittest.main()
